backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
pymongo==4.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
google-generativeai==0.3.0
python-dotenv==1.0.0